+++ Assets/Scripts/Character/Assets/AssetData.cs
@@ -2,5 +2,4 @@ using UnityEngine;
 
 public class AssetData : ScriptableObject
 {
-    [field: SerializeField] string AssetId { get; set; }
 }