+++ Assets/Scripts/Character/EntityManagers/EntitySpawner.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Security.Cryptography;
+using System.Threading.Tasks.Sources;
 using UnityEngine;
 using static EntitySpawner;
 
@@ -15,23 +16,50 @@ public class EntitySpawner : MonoBehaviour
         public EntityTypeEnum EntityType;
     }
 
+    public class SpawnedEntityInfo
+    {
+        public EntityTypeEnum EntityType;
+        public EntityInfo Info;
+        public EntityUI UIComponent;
+        public Character SpawnedEntityCharacter;
+    }
 
-    [field: SerializeField] EntityUI EntityUIPrefab;
-    [field: SerializeField] Character CharacterPrefab;
+
+    [field: SerializeField] EnemyEntityUI EntityUIPrefab;
+    [field: SerializeField] PlayerUI CharacterPrefab;
 
     [field: SerializeField] public Dictionary<string, SpawnedEntity> SpawnedEntities { get; private set; } = new Dictionary<string, SpawnedEntity>();
-    [field: SerializeField] public Dictionary<string, Character> Characters { get; private set; } = new Dictionary<string, Character>();
+    [field: SerializeField] public Dictionary<string, SpawnedEntityInfo> Characters { get; private set; } = new Dictionary<string, SpawnedEntityInfo>();
 
 
-    private ClientPacketHandler _clientPacketHandler;
+    private ClientCommunicationLayerManager _clientCommunicationLayerManager;
+
+    ClientCommunicationLayerManager ClientCommunicationLayerManager
+    {
+        get
+        {
+            if (_clientCommunicationLayerManager == null)
+            {
+                _clientCommunicationLayerManager = FindAnyObjectByType<ClientCommunicationLayerManager>();
+            }
+            return _clientCommunicationLayerManager;
+        }
+    }
     // Start is called before the first frame update
     private void OnEnable()
     {
-        if(_clientPacketHandler == null)
+        ClientCommunicationLayerManager.Handler.RegisterClientHandler<EntitySpawnPacket>(SpawnEntities);
+        ClientCommunicationLayerManager.Handler.RegisterClientHandler<PlayerEquipmentPacket>(UpdatePlayerEquipment);
+    }
+
+    private void UpdatePlayerEquipment(PlayerEquipmentPacket packet)
+    {
+        var playerUIComponent = Characters[packet.ClientId].UIComponent as PlayerUI;
+        if (playerUIComponent == null)
         {
-            _clientPacketHandler = FindAnyObjectByType<ClientPacketHandler>();
+            throw new Exception("Player UI Component is not of type PlayerUI");
         }
-        _clientPacketHandler.RegisterClientHandler<EntitySpawnPacket>(SpawnEntities);
+        playerUIComponent.UpdatePlayerBaseAsset(packet.PlayerBaseAsset);
     }
 
     private void SpawnEntities(EntitySpawnPacket packet)
@@ -93,11 +121,16 @@ public class EntitySpawner : MonoBehaviour
     private void SpawndEntity(string key, EntityInfo value)
     {
         var entity = Instantiate(EntityUIPrefab, value.Position, Quaternion.identity);
-        var ui = entity.GetComponent<EntityUI>();
-        ui.SpawnUI(value.EntityAssetId);
+        var ui = entity.GetComponent<EnemyEntityUI>();
+        var enemyInfo = value as EnemyEntityInfo;
+        if (enemyInfo == null)
+        {
+            throw new Exception("EntityInfo is not of type EnemyEntityInfo");
+        }
+        ui.SetEntityAssetId(enemyInfo.EntityAssetId);
         SpawnedEntity spawnedEntity = new SpawnedEntity()
         {
-            AssetId = value.EntityAssetId,
+            AssetId = enemyInfo.EntityAssetId,
             UIComponent = entity,
             EntityType = value.EntityType
         };
@@ -106,16 +139,25 @@ public class EntitySpawner : MonoBehaviour
 
     private void SpawnCharacter(string key, EntityInfo value)
     {
-        var character = Instantiate(CharacterPrefab, value.Position, Quaternion.identity);
-        Characters.Add(key, character);
+        var character = Instantiate(CharacterPrefab, value.Position, Quaternion.identity) ;
+        var spawnedCharacterInfo = new SpawnedEntityInfo()
+        {
+            EntityType = value.EntityType,
+            Info = value,
+            UIComponent = character,
+            SpawnedEntityCharacter = character.GetComponent<Character>() ?? throw new Exception("Character prefab does not have a Character component")
+        };
+        Characters.Add(key, spawnedCharacterInfo);
     }
 
     private void UpdateCharacter(string key, EntityInfo value)
     {
-        var character = Characters[key];
+        var playerEntityInfo = value as PlayerEntityInfo;
+        var character = Characters[key].SpawnedEntityCharacter;
         character.transform.position = value.Position;
         character.SetDirection(value.Rotation);
         character.SetMovement(value.Movement);
+        (Characters[key].UIComponent as PlayerUI).UpdatePlayerBaseAsset(playerEntityInfo.Equipment.BaseBody);
     }
 
     private void UpdateEntity(string key, EntityInfo value)
@@ -125,6 +167,6 @@ public class EntitySpawner : MonoBehaviour
 
     private void OnDisable()
     {
-        _clientPacketHandler.UnregisterClientHandler<EntitySpawnPacket>();
+        ClientCommunicationLayerManager.Handler.UnregisterClientHandler<EntitySpawnPacket>();
     }
 }